requirements.txt
fastapi==0.109.1
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# File handling
python-multipart==0.0.6

# Templates
jinja2==3.1.3

# Production (optional but recommended)
gunicorn==21.2.0